Creating seperate routes to make code look less complex.
