+++ routes/home.routes.js
@@ -1,0 +1,8 @@
+const express = require('express');
+const router = express.Router();
+
+router.get('/home', (req, res) => {
+    res.render('home')
+})  
+
+module.exports = router;
